Přidán parametrický middleweare a možnost porovnávat oprávnění pomocí AND nebo OR fixed #65 Přidáno ošetření getů pro editaci položek, kategorií a rolí na straně controllerů

+++ app/Http/Controllers/CategoryController.php
@@ -22,17 +22,13 @@ class CategoryController extends Controller
 
         echo "<script>console.log('Debug Objects: CategoryController' );</script>";
         Log::info('CategoryControler:showCategories');
-        if(Auth::permition()->return_verification == 1 || Auth::permition()->possibility_renting == 1 || Auth::permition()->edit_item == 1) {
 
-            $data = DB::table('categories')->leftJoin('items', 'categories.id', '=', 'items.categories')->select('categories.id', 'categories.name', 'categories.description', 'items.availability', DB::raw('COUNT(items.categories) as count'))->groupByRaw('categories.name, categories.description, items.availability, categories.id')->get();
-            $permition = DB::table('users')->join('permition', 'users.permition', '=', 'permition.id')->where('users.id', Auth::id())->select('permition.edit_item', 'permition.possibility_renting')->get();
 
-            //return $data    ;
-            return view('categories', ['categories' => $data, 'permition' => $permition]);
-        }
-        else{
-            return "0"; //sem dát view 403.blade
-        }
+        $data = DB::table('categories')->leftJoin('items', 'categories.id', '=', 'items.categories')->select('categories.id', 'categories.name', 'categories.description', 'items.availability', DB::raw('COUNT(items.categories) as count'))->groupByRaw('categories.name, categories.description, items.availability, categories.id')->get();
+        $permition = DB::table('users')->join('permition', 'users.permition', '=', 'permition.id')->where('users.id', Auth::id())->select('permition.edit_item', 'permition.possibility_renting')->get();
+
+        return view('categories', ['categories' => $data, 'permition' => $permition]);
+
 
 
     }
@@ -59,6 +55,11 @@ class CategoryController extends Controller
     {
         Log::info('CategoryControler:saveCategory');
 
+        if(Auth::permition()->edit_item != 1){
+            abort(403);
+            return;
+        }
+
         if($this->checkCategoryNameExist($request->categoryName) == "true" && $request->categoryName != $request->categoryNameOld){
             abort(409);
             return;
@@ -79,6 +80,11 @@ class CategoryController extends Controller
     {
         Log::info('CategoryControler:addNewCategory');
 
+        if(Auth::permition()->edit_item != 1){
+            abort(403);
+            return;
+        }
+
         $category = new categories;
         $category->name = 'Abecedně seřazená NOVÁ KATEGORIE';
         $check = $category->save();
@@ -90,6 +96,11 @@ class CategoryController extends Controller
     function removeCategory(Request $request)
     {
         Log::info('CategoryControler:removeCategory');
+
+        if(Auth::permition()->edit_item != 1){
+            abort(403);
+            return "0";
+        }
 
         $data = DB::table('items')->leftJoin('loans', 'items.id', '=', 'loans.item')->leftJoin('Users', 'loans.user', '=', 'Users.id')->Join('categories', 'items.categories', '=', 'categories.id')->orderBy('categories.name', 'asc')->orderBy('items.id', 'asc')->select('Users.id as userId', 'Users.name', 'Users.surname','categories.id as categoryId', 'categories.name as categoryName',  'items.id as itemId', 'items.name as itemName' , 'loans.id', 'loans.rent_from', 'loans.rent_to')->where('categories.id', $request->id)->get();
 
@@ -106,6 +117,11 @@ class CategoryController extends Controller
     function removeCategoryHard(Request $request)
     {
         Log::info('CategoryControler:removeCategoryHard');
+
+        if(Auth::permition()->edit_item != 1){
+            abort(403);
+            return;
+        }
 
         $loansController = new LoansController;
         $check1 = DB::table('loans')->Join('items','loans.item', '=', 'items.id')->where('items.categories', $request->categoryId)->delete();
